Fix interrupt vector functions, attempt to rewrite interrupt in assembly
For some reason, I'm facing major challenges with getting the address of the interrupt. I think I have a fundumental misunderstanding of how memory works in DOS, because while I'm easily able to obtain the address to my keyboard_interrupt function, it points to some random place in memory and the real location is much further ahead. I haven't been able to figure out why this is the case.